fix(ci): fix FreeRTOS mock headers — StackType_t not declared
FreeRTOS.h now includes portmacro.h first, which defines StackType_t
before task.h and Observable.hpp reference it.

+++ Tests/mocks/portmacro.h
@@ -1,8 +1,11 @@
 #pragma once
 #include <stdint.h>
-typedef unsigned long  StackType_t;
-typedef long           BaseType_t;
-typedef unsigned long  UBaseType_t;
-typedef uint32_t       TickType_t;
+
+/* Port-level types — must come before FreeRTOS.h redefines */
+typedef uint32_t      StackType_t;
+typedef long          BaseType_t;
+typedef unsigned long UBaseType_t;
+typedef uint32_t      TickType_t;
+
 #define portSTACK_TYPE  uint32_t
 #define portBASE_TYPE   long
